+++ R/03-tryCatchWithErrorsAndWarnings.R
@@ -3,14 +3,9 @@
 #' Catches errors and warnings and forwards them as an alert. If an error occurs, NULL is returned.
 #' If a warning occurs, the result is returned.
 #'
-#' @param expr expression to be evaluated.
-#' @param errorTitle (character) error message title.
-#' @param warningTitle (character) warning message title.
-#' @param alertStyle (character) Either "shinyjs", or "shinyalert". Specifies how an error or a
-#'  warning is given out. If "shinyjs" than shinyjs::alert is used; if "shinyalert" than
-#'  shinyalert::shinyalert is used.
 #' @param silent (logical) if TRUE prevents an alert, instead a warning will be displayed. Use this
 #'  when not applied within a Shiny session.
+#' @inheritParams shinyTools::shinyTryCatch
 #'
 #' @export
 tryCatchWithWarningsAndErrors <- function(expr,
@@ -18,78 +13,15 @@ tryCatchWithWarningsAndErrors <- function(expr,
                                           warningTitle = "",
                                           alertStyle = "shinyjs",
                                           silent = FALSE) {
-  tryCatchMessage <- list()
+  deprecate_warn("24.05.1",
+                 "DataTools::tryCatchWithWarningsAndErrors()",
+                 "shinyTools::shinyTryCatch()")
 
-  w.handler <- function(w) {
-    # warning handler
-    warningText <- w$message %>%
-      gsub(pattern = "\033\\[[0-9;]*m", replacement = "") %>% # removes ANSI codes (formatting of warnings)
-      paste0(collapse = "\n")
-    tryCatchMessage <<- c(tryCatchMessage, list(list(
-      text = warningText,
-      type = "warning"
-    )))
-    invokeRestart("muffleWarning")
-  }
-
-  e.handler <- function(e) {
-    # error handler
-    errorText <- e$message %>%
-      gsub(pattern = "\033\\[[0-9;]*m", replacement = "") %>% # removes ANSI codes (formatting of errors)
-      paste0(collapse = "\n")
-    tryCatchMessage <<- c(tryCatchMessage, list(list(
-      text = errorText,
-      type = "error"
-    )))
-    return(NULL)
-  }
-
-  res <- withCallingHandlers(tryCatch({
-    expr
-  },
-  error = e.handler),
-  warning = w.handler)
-
-  if (length(tryCatchMessage) == 0) {
-    # no error or warning occurred
-    return(res)
-  }
-
-  # extract type
-  allTypes <- sapply(tryCatchMessage, `[[`, "type")
-  if (any(allTypes == "error")) {
-    messageType <- "error"
-  } else {
-    messageType <- "warning"
-  }
-
-  # extract title
-  messageTitle <- switch(messageType, "error" = errorTitle, "warning" = warningTitle)
-
-  # extract text
-  allTexts <- sapply(tryCatchMessage, `[[`, "text")
-  messageText <- paste0(" ", allTypes, ": ", allTexts) %>%
-    as.character() %>%
-    paste0(collapse = "\n")
-
-  if (!silent) {
-    # give out alert within shiny app
-    switch(
-      alertStyle,
-      "shinyjs" = shinyjs::alert(paste(messageTitle, messageText, sep = "\n ")),
-      "shinyalert" = shinyalert::shinyalert(
-        title = messageTitle,
-        text = messageText,
-        type = messageType
-      )
-    )
-  } else {
-    # give out the error or warning, but always as a warning to not interrupt the calculation/app
-    warning(paste0(messageTitle, "\n", messageText), call. = FALSE)
-  }
-
-  # output result of expr
-  res
+  shinyTryCatch(expr = expr,
+                errorTitle = errorTitle,
+                warningTitle = warningTitle,
+                alertStyle = alertStyle,
+                inShiny = !silent)
 }
 
 # TEST MODULE -------------------------------------------------------------